Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scripthash-ignore-unsafeinline-expected.txt

ALERT: PASS (1/1)
CONSOLE MESSAGE: line 10: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
This tests that a valid hash value disables inline JavaScript, even if 'unsafe-inline' is present.
